backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]>=0.31.1
sqlmodel==0.0.14
asyncpg==0.29.0
pydantic[jinja2]==2.11.0
python-jose[cryptography]==3.3.0
# Use a newer pydantic that satisfies `mcp` and other packages
# pinned to a minimum compatible version
pydantic[email]>=2.11.0
pydantic-settings==2.5.2
python-multipart>=0.0.9
passlib[bcrypt]==1.7.4
# MCP and AI Agent SDKs
mcp==1.26.0
openai>=1.3.0
# Database and async support
sqlalchemy[asyncio]==2.0.23
# Environment configuration
python-dotenv==1.0.0
